+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Solidus 2.3.0 (master, unreleased)
 
+- Renamed bogus payment methods
+  `Spree::Gateway::BogusSimple` and `Spree::Gateway::Bogus` were renamed to `Spree::PaymentMethod::SimpleBogusCreditCard` and `Spree::PaymentMethod::BogusCreditCard`
+
 - Allow custom separator between a promotion's `base_code` and `suffix` [\#1951](https://github.com/solidusio/solidus/pull/1951) ([ericgross](https://github.com/ericgross))
 - Ignore `adjustment.finalized` on tax adjustments. [\#1936](https://github.com/solidusio/solidus/pull/1936) ([jordan-brough](https://github.com/jordan-brough))
 - Deprecate `#simple_current_order`